Supported refreshing the account attributes in the getAccountAttributes API.
If the account attribute was evaluated longer than a refresh cycle (e.g.
24h), the getAccountAttributes will recalulate the account attributes,
update the DB and return the fresh account attributes to the client.

This change also reserves a room for a future change to add backgroundtask that is running on ContactsProvider startup, which initialize the account attributes and store them in the DB.

Bug: 432299080
Bug: 434399211
Flag: android.provider.new_account_attributes_api_enabled
Test: manually test and unit test AccountAttributesManagerTest

Change-Id: I8d31dce2e27d13b996810fcbb8c9e9805df26608

	new file:   src/com/android/providers/contacts/AccountAttributesManager.java
	modified:   src/com/android/providers/contacts/ContactsProvider2.java
	new file:
tests/src/com/android/providers/contacts/AccountAttributesManagerTest.java

Change-Id: Ibbb947f4cf09fea4d90a5e29cba9ffdf25c0fba2